requirements.txt
openai
anthropic
google-generativeai
python-dotenv
httpx
